fix: resolve pre-existing clippy warnings in aurea-render
- Default for NodeId (missing Default impl)
- interaction: iterate map keys instead of (k,_) pair
- renderer: remove redundant as-u32 cast; allow too_many_arguments on blit_image_to_buffer
- cpu/blend: manual_checked_ops in blend_color_burn (checked_div + map_or)
- cpu/rasterizer: &mut Vec -> &mut [u32] on 8 private fns; allow too_many_arguments on draw_path/render_item
- cpu/scanline: allow too_many_arguments on fill_scanline; index loop to iterator-over-slice
- text/platform: collapse nested if-let with &&; allow too_many_arguments on render_text/blit_glyph
- text/directwrite_backend: allow arc_with_non_send_sync on COM FaceEntry
- text/fontdue_backend: direct return in file_stem; is_none_or; collapse nested ifs with let-chains
- text/mod: collapse nested if with let-chain in LruCache::insert

+++ aurea-render/src/renderer.rs
@@ -350,6 +350,7 @@ impl PlaceholderRenderer {
         }
     }
 
+    #[allow(clippy::too_many_arguments)]
     fn blit_image_to_buffer(
         image_data: &[u8],
         image_width: u32,
@@ -405,7 +406,7 @@ impl PlaceholderRenderer {
                     let dg = (dst >> 8) & 0xff;
                     let db = dst & 0xff;
                     let sa = a as u32;
-                    let inv_sa = (255 - sa) as u32;
+                    let inv_sa = 255 - sa;
                     let out_a = sa + (inv_sa * da) / 255;
                     if out_a == 0 {
                         buffer[buf_idx] = 0;
